Make order_by play well with other fields
The previous solution would clear all previous ``order_by`` clauses.

This commit, instead of manually controlling the ordering, adds an
option to the ``search`` method to just include the rank as an extra
field with a standardized name. As a result, the caller can use that
(standardized) name in the normal ``order_by`` method Django's default
QuerySet provides.