Add SearchQ object
Add a custom Q object subclass that can do a full-text search on a
column.

We only support the plaito_tsquery PostgreSQL function and using an
extra (undeclared to django) column that holds the pre-computed tsvector
data.

It is up to the user to create the column, index it as well as add a
trigger to update it.