Allow to add the FTS rank in the select clause
Add option to SearchQ object to add the rank in the select clause. This
is especially useful, when the caller wants to order the results by
similarity.

Since the aliases used in the SELECT clause columns are not remapped,
we can safely add the extra column to the select clause.